backend/requirements.txt
# Python==3.9
accelerate
huggingface-hub==0.24.7
joblib==1.4.2
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-experimental==0.3.2
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.137
language_data==1.2.0
multidict==6.1.0
multipledispatch==1.0.0
multitasking==0.0.11
neo4j==5.25.0
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
python-dotenv==1.0.1
sentence-transformers==3.2.1
torch
transformers==4.44.2
decorator
json-repair
tiktoken
fastapi
uvicorn
psycopg2-binary==2.9.10
python-magic
pymupdf4llm
python-multipart
passlib
bcrypt==4.0.0
python-jose[cryptography]
python-multipart
email-validator